perf(bundle): lazy cascade-simulator + buildRiskCards extracted + estimators barrel bypass
Three more bundle-shape wins on the critical path:

1. `cascade-simulator` (446 LOC) was eagerly imported by useApexStore
   but only used in three fire-and-forget replay actions. Wrapped
   each call site in a `void loadSimulateCascadeAsync().then(...)`
   that dynamic-imports the module. Replay is opt-in (user clicks
   REPLAY), so the sim chunk now defers until then. Also dropped
   a dead `simulateCascade` sync import that was never called.

2. `buildRiskCards` is a pure function but lived in graph-data.ts.
   `RiskPropagationFlow` (on critical path) only needed this helper
   but transitively pulled the 3000-line dataset. Extracted to a
   new 27-LOC `src/lib/risk-cards.ts`; graph-data no longer needs
   its `CausalShock` / `RiskPropagationCard` type imports either.

3. `StructuralMetrics` (critical path) was importing
   `omegaBridgeDensity` through the `@/lib/estimators` barrel,
   which re-exports from 9 modules (~2500 LOC: bocpd, chi-star,
   cvar-w1, lppls-fit, moran, nlme, ph-fit, transfer-entropy,
   persistent-homology). Webpack's barrel tree-shaking is unreliable
   when modules carry side effects; safer to direct-import the
   76-LOC leaf. Same surface API, much smaller dep graph.

vitest 1524/1524 pass. tsc clean on all touched files.
