+++ src/components/RiskPropagationFlow.tsx
@@ -5,7 +5,7 @@ import { motion, AnimatePresence } from "framer-motion";
 import { useApexStore } from "@/stores/useApexStore";
 import { useFilteredGraph } from "@/hooks/useFilteredGraph";
 import { getDomainColor } from "@/lib/graph-color";
-import { buildRiskCards } from "@/lib/graph-data";
+import { buildRiskCards } from "@/lib/risk-cards";
 import { useTemporalGraph } from "@/hooks/useTemporalGraph";
 import type { NodeTemporalState } from "@/lib/temporal-data";
 import { getNodeDataDescription } from "@/lib/real-timeseries";
